web/requirements/base.txt
commonmark
dateutils
six
pandas
django
django-ckeditor
django-crispy-forms
django-datatrans
django-debug-toolbar
django-extensions
django-grappelli
django-model-utils
django-modeltranslation
# django-pandas
django-rest-swagger
# django-reversion
django-tables2
djangorestframework
fuzzywuzzy
gunicorn
httplib2
ipython
jinja2
numpy
pillow
# pypdf2
python-dateutil
python-memcached
pytz
pyyaml
# qrcode
redis
selenium
setuptools
simplejson
sorl-thumbnail
#sphinx
unicodecsv
unidecode
xlsxwriter
psycopg2
ghostscript
uwsgi
